requirements.txt
torch==1.7.0
torchvision==0.8.1
fastapi==0.61.1
opencv-python==4.3.0.36
Pillow==6.2.1
uvicorn==0.11.8
requests==2.23.0
numpy==1.19.5